load_img: memory start might be out of 32-bit address

+++ include/memory/image_loader.h
@@ -6,6 +6,6 @@
 
 long load_gz_img(const char *filename);
 
-long load_img(char* img_name, char *which_img, unsigned load_start, size_t img_size);
+long load_img(char* img_name, char *which_img, uint64_t load_start, size_t img_size);
 
 #endif //  __IMAGE_LOADER_H__
